+++ NEWS.md
@@ -5,6 +5,8 @@ haven format attributes.
 * Added `fapply2()` function to format and combine two variables.
 * Added `na` and `zero` parameters to `fmt_cnt_pct()` to apply specified string 
 in cases where the count is NA or zero.  
+* Removed `labels.data.frame()` function from this package, as it has now
+been moved to the **common** package.
 
 # fmtr 1.5.7